Moved some barcode types to own class
